+++ src/Queries.Core/Parts/ITable.cs
@@ -1,5 +1,8 @@
 namespace Queries.Core.Parts
 {
+    /// <summary>
+    /// Marker interface for elements that can be used where a table 
+    /// </summary>
     public interface ITable
     {
         /// <summary>